♻️ refactor(model-runtime): resolve circular dependencies and migrate to node environment
- Migrate test environment from edge-runtime to node across 29 test files
- Move fetch and parseModels utilities from packages/utils to src/utils
- Convert package self-imports to relative imports to prevent circular dependencies
- Update import paths in genServerAiProviderConfig to use correct module location
