+++ packages/model-runtime/src/providers/anthropic/generateObject.test.ts
@@ -1,4 +1,4 @@
-// @vitest-environment edge-runtime
+// @vitest-environment node
 import { describe, expect, it, vi } from 'vitest';
 
 import { createAnthropicGenerateObject } from './generateObject';
