+++ packages/model-runtime/src/providers/bedrock/index.test.ts
@@ -3,10 +3,10 @@ import {
   InvokeModelCommand,
   InvokeModelWithResponseStreamCommand,
 } from '@aws-sdk/client-bedrock-runtime';
-import { AgentRuntimeErrorType } from '@lobechat/model-runtime';
 import { ModelProvider } from 'model-bank';
 import { Mock, afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
+import { AgentRuntimeErrorType } from '../../types/error';
 import * as debugStreamModule from '../../utils/debugStream';
 import { LobeBedrockAI, experimental_buildLlama2Prompt } from './index';
