+++ packages/model-runtime/src/providers/qwen/createImage.test.ts
@@ -1,4 +1,4 @@
-// @vitest-environment edge-runtime
+// @vitest-environment node
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
 import { CreateImageOptions } from '../../core/openaiCompatibleFactory';
