+++ packages/model-runtime/src/providers/taichu/index.test.ts
@@ -1,9 +1,9 @@
 // @vitest-environment node
-import { LobeOpenAICompatibleRuntime } from '@lobechat/model-runtime';
 import { ModelProvider } from 'model-bank';
 import OpenAI from 'openai';
 import { Mock, afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
+import { LobeOpenAICompatibleRuntime } from '../../core/BaseAI';
 import { testProvider } from '../../providerTestUtils';
 import * as debugStreamModule from '../../utils/debugStream';
 import { LobeTaichuAI } from './index';
