+++ packages/utils/src/index.ts
@@ -4,9 +4,9 @@ export * from './detectChinese';
 export * from './format';
 export * from './imageToBase64';
 export * from './keyboard';
+export * from './merge';
 export * from './number';
 export * from './object';
-export * from './parseModels';
 export * from './pricing';
 export * from './safeParseJSON';
 export * from './sleep';
